pyproject.toml
[project]
name = "parse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bs4>=0.0.2",
    "pyopenssl>=25.1.0",
    "requests>=2.32.4",
    "ruff>=0.11.13",
    "selenium>=4.33.0",
    "sqlalchemy>=2.0.41",
]
